% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene Out Filter Func.R
\name{filter.out.genes}
\alias{filter.out.genes}
\title{Filter out genes}
\usage{
\method{filter}{out.genes}(data.table, gene.list, by.rownames = TRUE,
  col = NULL)
}
\arguments{
\item{data.table}{Data.frame with rownames as genes.}

\item{gene.list}{A vector with gene names or name to be filtred out}

\item{by.rownames}{default filters by row names of the data.table. If false must provide col index.}

\item{col}{Index of column postion if rownames are not to be used. Must provide col index if by.row is false}
}
\value{
Returns a new data frame with genes matching to genes.list filtered out.
}
\description{
Filters out genes non-specficically from a data table
}
\examples{
filter.out.genes(expressed.genes, c("RPL11", "RPS", by.rownames = TRUE))
filter.out.genes(expressed.genes, c("RP"), byrownames = FALSE, col = 2)
}
