% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter plot.R
\name{gene.scatter}
\alias{gene.scatter}
\title{Plot sample vs sample}
\usage{
gene.scatter(data.set, x.sample, y.sample, pheno.table = NULL,
  names.col = NULL, text.transparency = 0.5,
  point.transparency = 0.1, min.cutoff = log(0.101), diff.cutoff = 2)
}
\description{
Plot sample vs sample
}
