% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergetable.R
\name{mergeTables}
\alias{mergeTables}
\title{Merge Tables}
\usage{
mergeTables(wd, commonName, colsToMerge)
}
\arguments{
\item{wd}{Working directory in which all folders may be viewed}

\item{commonName}{A single string representing the exact name all files share}

\item{colsToMerge}{integers representing the cols to keep}
}
\value{
Returns a new data.table which has combined all the tables into one large table
}
\description{
Merges tables in a directory with a common name
}
\examples{
main.table <- mergeTables(wd = "C:/Users/brendan/Documents/sravandevanathan/ballgown",
commonName = "gene_abundance.tab", 
colsToMerge = c(2,5,6,8))

main.table <- mergeTables(wd = "C:/Users/brendan/Documents/sravandevanathan/ballgown",
commonName = "t_data.ctab", 
colsToMerge = c(4,5,6,10,12))
}
