% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Var across samples.R
\name{Var.samples}
\alias{Var.samples}
\title{Gene Variation Between Rows}
\usage{
Var.samples(data.set, gene.list, graph = FALSE, pretty.names = FALSE,
  func = log2, zero = 0.01)
}
\arguments{
\item{data.set}{Data.frame with rownames as genes and cols as different samples.}

\item{gene.list}{A vector with gene names or name to be filtred}

\item{graph}{a bolean stating if a simple bar graph should be returned}

\item{pretty.names}{a bolean stating if gene names should returned without unique start and stop identifier:
may create duplicates. In such a case a pretty col will be returned instead. See pretty.gene.names func.}

\item{func}{which log scale should the data be  subjected to}

\item{zero}{addition to offset taking log of zero. Otherwise zeros in the dataset will create inf that mutate entire roww}
}
\value{
Returns a new data frame with only genes matching to genes.list and a variation column.
If a gene.list name fails to match it will be stated. See gene.filter func. 
If pretty.names is TRUE and duplicates exist a pretty gene col will be return in the data.frame
}
\description{
Filters genes non-specficically from a data table and calculates variation between samples. Returns a dataset
with a var col.
}
\examples{
filtered.data <- Var.samples(expressed.genes, gene.list = c("RPL11"))
filtered.data <- Var.samples(expressed.genes, gene.list = c("RPL11"), pretty.names = TRUE, graph = TRUE)
}
